feat(auth): implement user authentication with signup, signin, and profile retrieval
feat(workout): restrict workout access to authorized users and implement user-specific workout retrieval

feat(workoutExercise): add user-specific workout exercise retrieval and authorization checks

refactor(setLog): enhance set log retrieval and authorization based on user ID

refactor: organize routes and middleware for better structure and clarity

+++ src/modules/SetLog/setLog.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from "express";
-import { ErrorHandler, NotFoundError } from "../../Helpers/ErrorHandler.js";
+import { ErrorHandler, NotFoundError, UnauthorizedError } from "../../Helpers/ErrorHandler.js";
 import { ResponseHandler } from "../../Helpers/ResponseHandler.js";
 import { SetlogDao } from "./setLog.dao.js";
 import { CreateSetLogSchema, idSchema, UpdateSetLogSchema } from "./setLog.dtos.js";
@@ -12,9 +12,10 @@ export class SetLogController {
     constructor() {
         this.dao = new SetlogDao();
     }
-    async getAll(req: Request, res: Response) {
+    async getAllByUserId(req: Request, res: Response) {
         try {
-            const setLogs = await this.dao.getAll();
+            const userId = (req as any).user.id;
+            const setLogs = await this.dao.getAllByUserId(userId);
             return ResponseHandler.success(res, setLogs);
         } catch (error) {
             return ErrorHandler.handle(error, res);
@@ -23,10 +24,14 @@ export class SetLogController {
 
     async getOne(req: Request, res: Response) {
         try {
+            const userId = (req as any).user.id;
             const id = idSchema.parse(req.params.id);
             const setLog = await this.dao.getOne(id);
             if (!setLog) {
-                throw new NotFoundError('Set log not found');
+                throw new NotFoundError();
+            }
+            if (setLog.workoutExercise.workout.userId !== userId) {
+                throw new UnauthorizedError();
             }
             return ResponseHandler.success(res, setLog);
         } catch (error) {
@@ -46,8 +51,16 @@ export class SetLogController {
 
     async update(req: Request, res: Response) {
         try {
+            const userId = (req as any).user.id;
             const id = idSchema.parse(req.params.id);
             const data = UpdateSetLogSchema.parse(req.body); 
+            const setFound = await this.dao.getOne(id);
+            if (!setFound) {
+                throw new NotFoundError();
+            }
+            if (setFound.workoutExercise.workout.userId !== userId) {
+                throw new UnauthorizedError();
+            }
             const setLog = await this.dao.update(id, data);
             return ResponseHandler.updated(res, setLog);
         } catch (error) {
@@ -57,7 +70,15 @@ export class SetLogController {
 
     async delete(req: Request, res: Response) {
         try {
+            const userId = (req as any).user.id;
             const id = idSchema.parse(req.params.id);
+            const setFound = await this.dao.getOne(id);
+            if (!setFound) {
+                throw new NotFoundError();
+            }
+            if (setFound.workoutExercise.workout.userId !== userId) {
+                throw new UnauthorizedError();
+            }
             await this.dao.delete(id);
             return ResponseHandler.success(res, null, 'Set log deleted successfully');
         } catch (error) {
